create of subpackage and move interfaces to it
