--- src/main/java/com/klodnicki/Bike/v2/service/interfacee/AddService.java
@@ -1,6 +1,0 @@
-package com.klodnicki.Bike.v2.service.interfacee;
-
-public interface AddService<T, G> {
-
-    T add(G obj);
-}
