+++ src/main/java/com/klodnicki/Bike/v2/service/interfacee/GenericBikeService.java
@@ -3,6 +3,7 @@ package com.klodnicki.Bike.v2.service.interfacee;
 import com.klodnicki.Bike.v2.DTO.bike.BikeForAdminResponseDTO;
 import com.klodnicki.Bike.v2.DTO.bike.BikeRequestDTO;
 import com.klodnicki.Bike.v2.model.entity.Bike;
+import com.klodnicki.Bike.v2.service.interfacee.basic.operation.*;
 
 import java.util.List;
