--- src/main/java/com/klodnicki/Bike/v2/service/interfacee/SaveService.java
@@ -1,5 +1,0 @@
-package com.klodnicki.Bike.v2.service.interfacee;
-
-public interface SaveService<T> {
-    T save(T obj);
-}
