+++ src/main/java/com/klodnicki/Bike/v2/service/interfacee/basic/operation/AddService.java
@@ -1,0 +1,6 @@
+package com.klodnicki.Bike.v2.service.interfacee.basic.operation;
+
+public interface AddService<T, G> {
+
+    T add(G obj);
+}
