+++ src/main/java/com/klodnicki/Bike/v2/service/interfacee/basic/operation/FindService.java
@@ -1,4 +1,4 @@
-package com.klodnicki.Bike.v2.service.interfacee;
+package com.klodnicki.Bike.v2.service.interfacee.basic.operation;
 
 import java.util.List;
