+++ src/main/java/com/klodnicki/Bike/v2/service/interfacee/basic/operation/SaveService.java
@@ -1,0 +1,5 @@
+package com.klodnicki.Bike.v2.service.interfacee.basic.operation;
+
+public interface SaveService<T> {
+    T save(T obj);
+}
